+++ traficoTiming/TraficoTiming.Services/Models/UpdateRawTimingResultModel.cs
@@ -3,6 +3,6 @@ namespace TraficoTiming.Services.Models;
 public class UpdateRawTimingResultModel
 {
     public Guid ResultId { get; set; }
-    public double? AdjustedTimeSeconds { get; set; }
+    public decimal? AdjustedTimeSeconds { get; set; }
     public string? ReviewerNote { get; set; }
 }
